Drop reference to mqtt client on disconnect
Without this change, a circular reference will continue to exist between the
Device object and the mqtt client object after disconnection.  This shouldn't
be a problem since Python can GC cyclic isolates, but it appears the __del__
method of the mqtt client may be causing that to fail (e.g. by adding an
external reference via a callback).

In any case, after adding this, unused libdyson Device and associated mqtt
client objects seem to be cleaned up properly in my use-case.

We can also call disconnect() in case of exceptions to clear up the client
state and avoid explicit calls to loop_stop.

Fixes shenxn/ha-dyson#34